requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==1.5.7
azure-common==1.1.28
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
billiard==3.6.4.0
boto3==1.17.69
botocore==1.20.112
celery==5.2.5
certifi==2024.2.2
charset-normalizer==2.0.4
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmdstanpy==1.2.0
contourpy==1.2.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
cron-descriptor==1.4.0
cryptography==3.4.8
cycler==0.12.1
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20230430
distlib==0.3.4
dj-database-url==0.5.0
Django==4.1.6
django-allauth==0.58.2
django-celery-beat==2.5.0
django-celery-results==2.4.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-crum==0.7.9
django-environ==0.4.5
django-extensions==3.1.2
django-filter==21.1
django-heroku==0.0.0
django-htmx==1.15.0
django-import-export==3.3.3
django-nine==0.2.5
django-phonenumber-field==6.0.0
django-preventconcurrentlogins==0.8.2
django-resized==0.3.11
django-rest-swagger==2.2.0
django-simple-history==2.12.0
django-storages==1.11.1
django-tenant-users==1.1.1
django-tenants==3.4.7
django-timezone-field==6.1.0
djangorestframework==3.13.1
djongo==1.3.6
drf-writable-nested==0.6.3
drf-yasg==1.20.0
elastic-transport==8.0.1
elasticsearch==7.17.1
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
filelock==3.6.0
fonttools==4.45.1
fuzzywuzzy==0.18.0
gunicorn==20.1.0
holidays==0.37
idna==3.2
importlib-metadata==4.11.1
importlib-resources==6.1.1
inflect==7.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.7.3
jmespath==0.10.0
jsonschema==2.4.0
kiwisolver==1.4.5
kombu==5.2.4
lml==0.1.0
Markdown==2.6.10
markup==0.2
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib==3.8.2
mongoengine==0.24.0
ndg-httpsclient==0.5.1
numpy==1.26.2
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==21.3
pandas==1.5.3
pdf2image==1.16.3
platformdirs==2.5.1
plotly==5.18.0
prompt-toolkit==3.0.28
prophet==1.1.5
psycopg2-binary==2.9.1
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.14.1
pydantic==2.5.2
pydantic_core==2.14.5
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
Pygments==2.2.0
pyOpenSSL==20.0.1
pyparsing==3.0.7
PyPDF2==3.0.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2022.1
PyYAML
rave-python==1.2.15
redis==4.2.2
requests==2.26.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
simplejson==3.17.6
six==1.16.0
slippers==0.6.2
sqlparse==0.2.4
stanio==0.3.0
tablib==3.5.0
telesign==2.2.2
tenacity==8.2.3
thefuzz==0.19.0
toml==0.10.2
tqdm==4.66.1
twilio==6.52.0
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.6
vine==5.0.0
virtualenv==20.14.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.14.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.7.0
